feat: Introduce coroutine dispatchers for use cases
This commit introduces a `DispatcherModule` to provide `Default`, `IO`, and `Main` dispatchers via Hilt.

The following use cases have been updated to utilize the `DefaultDispatcher` for their operations:
- `FetchAlbumListUseCase`
- `GetAlbumsUseCase`
- `IsDatabasePopulatedUseCase`
- `IsOnlineUseCase`
- `SaveAlbumsUseCase`

Unit tests for these use cases have also been updated to provide an `UnconfinedTestDispatcher`.
The `AlbumsUiStateMapperTest` was updated to assert the `isOnline` property in the `Loaded` state.